use-cases/azure-openai-pinecone/requirements.txt
pinecone>=5.4.0
openai>=1.60.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
numpy>=1.26.0
pypdf>=5.4.0
python-docx>=1.1.0
